+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # SensorLogger
 
 ### 0.1.0 UNRELEASED
+  * Change Database Scheme set user_id and uuid unique by migration
+  * Modified DeviceId to unique for each user
+  * Added more data validation on createLog #69
   * Added some more data validation #69 @vitoller
   * Modified registerDevice set deviceType optional
   * Modified registerDevice set deviceGroup optional